+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the Keyboard Macro Bata extension will be documented in t
 
 ### [Unreleased]
 - Fix
+  - Complex snippet insertion with multiple occurrences of a single placeholder.
   - Failure to detect typed characters under some conditions.
   - Wrong location of multi-cursor after playback of typing under some conditions.
   - Incorrect prediction of multi-cursor location on typing detection under some conditions.